Organize classes | Prefer Dictionary<string, object> for parameters
